+++ libs/viewer/src/Settings_generated.cpp
@@ -800,6 +800,8 @@ int parse(jsmntok_t const* tokens, int i, const char* jsonChunk, VsmShadowOption
             i = parse(tokens, i + 1, jsonChunk, &out->anisotropy);
         } else if (compare(tok, jsonChunk, "mipmapping") == 0) {
             i = parse(tokens, i + 1, jsonChunk, &out->mipmapping);
+        } else if (compare(tok, jsonChunk, "highPrecision") == 0) {
+            i = parse(tokens, i + 1, jsonChunk, &out->highPrecision);
         } else if (compare(tok, jsonChunk, "minVarianceScale") == 0) {
             i = parse(tokens, i + 1, jsonChunk, &out->minVarianceScale);
         } else if (compare(tok, jsonChunk, "lightBleedReduction") == 0) {
@@ -820,6 +822,7 @@ std::ostream& operator<<(std::ostream& out, const VsmShadowOptions& in) {
     return out << "{\n"
         << "\"anisotropy\": " << int(in.anisotropy) << ",\n"
         << "\"mipmapping\": " << to_string(in.mipmapping) << ",\n"
+        << "\"highPrecision\": " << to_string(in.highPrecision) << ",\n"
         << "\"minVarianceScale\": " << (in.minVarianceScale) << ",\n"
         << "\"lightBleedReduction\": " << (in.lightBleedReduction) << "\n"
         << "}";