+++ shaders/src/depth_main.fs
@@ -40,12 +40,12 @@ void main() {
 
 #if defined(VARIANT_HAS_VSM)
     // interpolated depth is stored in vertex_worldPosition.w (see main.vs)
+    // we always compute the "negative" side of ELVSM because the cost is small, and this allows
+    // EVSM/ELVSM choice to be done on the CPU side more easily.
     highp float depth = vertex_worldPosition.w;
-    depth = exp(depth);
+    depth = exp(frameUniforms.vsmExponent * depth);
     fragColor.xy = computeDepthMomentsVSM(depth);
-    fragColor.zw = vec2(0.0);
-    // enable for full EVSM (needed for large blurs). RGBA16F needed.
-    //fragColor.zw = computeDepthMomentsVSM(-1.0/depth);
+    fragColor.zw = computeDepthMomentsVSM(-1.0 / depth); // requires at least RGBA16F
 #elif defined(VARIANT_HAS_PICKING)
     outPicking.x = getObjectUniforms().objectId;
     outPicking.y = floatBitsToUint(vertex_position.z / vertex_position.w);